pyproject.toml
[tool.poetry]
name = "tonystock"
version = "0.1.0"
description = "Stock analysis project"
authors = ["Jeff"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
flask = "^3.1.0"
uvicorn = "^0.29.0"
pandas = "^2.2.3"
numpy = "^1.24.4"
httpx = "^0.27.2"
aiohttp = "^3.11.11"
beautifulsoup4 = "^4.12.3"
crawl4ai = "^0.4.247"
anthropic = "^0.45.2"
openai = "^1.60.2"
tiktoken = "^0.8.0"
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.7.1"
tenacity = "^9.0.0"
opik = "^1.4.5"
typing-extensions = "^4.12.2"
rich = "^13.9.4"
litellm = "^1.59.8"
pymongo = "^4.10.1"
redis = "^5.2.1"
playwright = "^1.49.1"
pdfminer-six = "^20240706"
yfinance = "^0.2.52"
ta-lib = "^0.6.1"
pytest = "^8.3.4"
pillow = "^10.4.0"
openpyxl = "^3.1.5"
python-pptx = "^1.0.2"
markdownify = "^0.14.1"
speechrecognition = "^3.14.1"
pydub = "^0.25.1"
mysql-connector-python = "^9.2.0"
sentry-sdk = "^2.20.0"
youtube-transcript-api = "^0.6.2"
docling = "^2.18.0"
opencc = "^1.1.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
